# swyftr-swagger.yaml
swagger: '2.0'
info:
  version: '1'
  title: Swyftr
  description: Backend API for Swyftr system
host: 52.221.58.5
basePath: /api/v1
schemes:
  - https
...
paths:
...
  /order:
    post:
      summary: Order Creation
      tags:
        - Order
        - Sprint2
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        '200':
          description: OK
          schema:
            properties:
              orderId:
                type: string
        '401':
          description: Bad Request
        '404':
          description: Existing Order Not Found When Updating
        '500':
          description: Internal Server Error
  /order/{orderId}:
    get:
      summary: Get Order Details By Order Id
      tags:
        - Order
        - Sprint2
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: orderId
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Order'
        '401':
          description: Bad Request
        '404':
          description: Existing Order Not Found When Updating
        '500':
          description: Internal Server Error
    delete:
      summary: Remove Order
      tags:
        - Order
        - Sprint2
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: orderId
          required: true
          type: number
      responses:
        '200':
          description: OK
        '500':
          description: Fail
  /order/customer/{customerId}:
    get:
      summary: Get Order Details By Customer Id
      tags:
        - Order
        - Sprint2
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: customerId
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Order'
        '401':
          description: Bad Request
        '404':
          description: Existing Order Not Found When Updating
        '500':
          description: Internal Server Error
  /order/change_status_qr:
    post:
      summary: Change status of an order via QR by Customer
      tags:
        - Partner
        - Sprint2
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              qrString:
                type: string
              status:
                type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
        '400':
          description: Bad Request
        '500':
          description: Fail
  /order/sys/all:
    get:
      summary: Get all orders, (can get by status)
      tags:
        - Order
        - Sprint2
        - System
      produces:
        - application/json
      parameters:
        - in: query
          required: false
          name: status
          type: string
          enum:
            - INIT
            - ACCP
            - DISPATCH
            - DELI
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              data:
                type: array
                items:
                  $ref: '#/definitions/Order'
        404:
          description: Bad Request
  /order/{orderId}/rider:
      get:
        summary: Get rider information by order id
        tags:
          - Order
          - Sprint3
          - Customer
        produces:
          - application/json
        parameters:
          - in: path
            required: true
            name: orderId
            type: number
        responses:
          500:
            description: SERVER ERROR
          404:
            description: BAR REQUEST
          200:
            description: OK
            schema:
              type: object
              properties:
                status:
                  type: integer
                rider:
                  type: object
                  properties:
                    firstName:
                      type: string
                    lastName:
                      type: string
                    mobile:
                      type: string
                    vehicleNo:
                      type: string
                    riderId:
                      type: number
  /order/sys/{orderId}:
    get:
      summary: Get Order Details By Order Id
      tags:
        - Order
        - Sprint2
        - Customer
      produces:
        - application/json
      parameters:
        - in: path
          name: orderId
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Order'
        '401':
          description: Bad Request
        '404':
          description: Existing Order Not Found When Updating
        '500':
          description: Internal Server Error
  /order/sys/change_status/{orderId}:
    post:
      summary: Change status of an order
      tags:
        - Partner
        - Sprint2
        - System
      produces:
        - application/json
      parameters:
        - in: path
          name: orderId
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              status:
                type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
        '400':
          description: Bad Request
        '500':
          description: Fail
  /order/sys/change_status_qr:
    post:
      summary: Change status of an order via QR by Rider
      tags:
        - Partner
        - Sprint2
        - System
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              qrString:
                type: string
              status:
                type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
        '400':
          description: Bad Request
        '500':
          description: Fail
...
  /report/orders/{status}/{fromDate}/{toDate}/{orgId}:
    get:
      summary: Orders By Status By Partner
      tags:
        - Sprint2
        - Reporting
        - Partner
      produces:
        - application/json
      parameters:
        - in: path
          name: status
          type: string
          required: true
        - in: path
          name: fromDate
          type: string
          required: true
        - in: path
          name: toDate
          type: string
          required: true
        - in: path
          name: orgId
          type: integer
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              orders:
                type: array
                items:
                  $ref: '#/definitions/Order_Status_Response'
...
  /paymentInfo/cus:
    post:
      summary: Add Payment Info
      tags:
        - Sprint2
        - Payment
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/PaymentInfo'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              paymentInfoId:
                type: integer
  /paymentInfo/sys:
    post:
      summary: Add Payment Info
      tags:
        - Sprint2
        - Payment
        - Customer
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/PaymentInfo'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              status:
                type: integer
              paymentInfoId:
                type: integer
...
definitions:
...
  Order:
    type: object
    properties:
      orgId:
        type: number
      branchId:
        type: number
      customer:
        type: number
      orderAddress:
        type: string
      deliveryRequired:
        type: number
      orderId:
        type: number
      status:
        type: string
        enum: [INIT, ACCP, DISPATCH, DELI]
      latitude:
        type: number
      longitude:
        type: number
      deliveryCost:
        type: number
      deliveryDate:
        type: number
      deliveryETA:
        type: string
      qr:
        type: string
      deliveryContactNumber:
        type: string
      writtenAddress:
        type: string
      paymentStatus:
        type: string
        enum: [CARD, IPG]
      items:
        type: array
        items:
          $ref: '#/definitions/OrderItem'
      friendReceive:
        $ref: '#/definitions/FriendReceive'
...
  OrderItem:
    type: object
    properties:
      branchStockId:
        type: number
      quantity:
        type: integer
...
  FriendReceive:
    type: object
    properties:
      name:
        type: string
      address:
        type: string
      phone:
        type: number
...
  Order_Status_Response:
    type: object
    properties:
      orgId:
        type: integer
      name:
        type: string
      total:
        type: number
      branchId:
        type: integer
      branchName:
        type: string
...
  PaymentInfo:
    type: object
    properties:
      orderId:
        type: integer
      transactionId:
        type: integer
      totalAmount:
        type: number
      email:
        type: string
      contactNumber:
        type: string
      deliveryAddress:
        type: string
      status:
        type: string
...